+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/DeploymentServiceImpl.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.ui.process.deployment
 
+import akka.actor.ActorSystem
 import com.typesafe.scalalogging.LazyLogging
 import db.util.DBIOActionInstances.DB
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionType.ProcessActionType
@@ -18,9 +19,10 @@ import pl.touk.nussknacker.ui.process.deployment.LoggedUserConversions.LoggedUse
 import pl.touk.nussknacker.ui.process.exception.{DeployingInvalidScenarioError, ProcessIllegalAction}
 import pl.touk.nussknacker.ui.process.repository.FetchingProcessRepository.FetchProcessesDetailsQuery
 import pl.touk.nussknacker.ui.process.repository.ProcessDBQueryRepository.ProcessNotFoundError
-import pl.touk.nussknacker.ui.process.repository.{DbProcessActionRepository, DeploymentComment, FetchingProcessRepository, ProcessDBQueryRepository, DBIOActionRunner}
+import pl.touk.nussknacker.ui.process.repository.{DBIOActionRunner, DbProcessActionRepository, DeploymentComment, FetchingProcessRepository, ProcessDBQueryRepository}
 import pl.touk.nussknacker.ui.security.api.{LoggedUser, NussknackerInternalUser}
 import pl.touk.nussknacker.ui.validation.ProcessValidation
+import pl.touk.nussknacker.ui.util.FutureUtils._
 import slick.dbio.DBIOAction
 
 import java.time.Clock
@@ -40,7 +42,9 @@ class DeploymentServiceImpl(dispatcher: DeploymentManagerDispatcher,
                             processValidation: ProcessValidation,
                             scenarioResolver: ScenarioResolver,
                             processChangeListener: ProcessChangeListener,
-                            clock: Clock = Clock.systemUTC()) extends DeploymentService with LazyLogging {
+                            scenarioStateTimeout: Option[FiniteDuration],
+                            clock: Clock = Clock.systemUTC())
+                           (implicit system: ActorSystem) extends DeploymentService with LazyLogging {
 
   def getDeployedScenarios(processingType: ProcessingType)
                           (implicit ec: ExecutionContext): Future[List[DeployedScenarioData]] = {
@@ -74,7 +78,7 @@ class DeploymentServiceImpl(dispatcher: DeploymentManagerDispatcher,
     checkCanPerformActionAndAddInProgressAction[Unit](processId.id, actionType, _.lastDeployedAction.map(_.processVersionId), _ => None).flatMap {
       case (processDetails, actionId, versionOnWhichActionIsDone, buildInfoProcessIngType) =>
         runDeploymentActionWithNotifications(actionType, actionId, processId, versionOnWhichActionIsDone, deploymentComment, buildInfoProcessIngType) {
-          dispatcher.deploymentManager(processDetails.processingType).cancel(processId.name, user.toManagerUser)
+          dispatcher.deploymentManagerUnsafe(processDetails.processingType).cancel(processId.name, user.toManagerUser)
         }
     }
   }
@@ -96,7 +100,7 @@ class DeploymentServiceImpl(dispatcher: DeploymentManagerDispatcher,
           case Success(validationResult) =>
             //we notify of deployment finish/fail only if initial validation succeeded
             val deploymentFuture = runDeploymentActionWithNotifications(actionType, actionId, processIdWithName, versionOnWhichActionIsDone, deploymentComment, buildInfoProcessIngType) {
-              dispatcher.deploymentManager(processDetails.processingType)
+              dispatcher.deploymentManagerUnsafe(processDetails.processingType)
                 .deploy(validationResult.processVersion, validationResult.deploymentData, validationResult.resolvedScenario, savepointPath)
             }
             Future.successful(deploymentFuture)
@@ -107,7 +111,7 @@ class DeploymentServiceImpl(dispatcher: DeploymentManagerDispatcher,
   protected def validateBeforeDeploy(processDetails: BaseProcessDetails[CanonicalProcess], actionId: ProcessActionId)
                                     (implicit user: LoggedUser, ec: ExecutionContext): Future[DeployedScenarioData] = {
     validateProcess(processDetails)
-    val deploymentManager = dispatcher.deploymentManager(processDetails.processingType)
+    val deploymentManager = dispatcher.deploymentManagerUnsafe(processDetails.processingType)
     for {
       // TODO: scenario was already resolved during validation - use it here
       resolvedCanonicalProcess <- Future.fromTry(scenarioResolver.resolveScenario(processDetails.json, processDetails.processCategory))
@@ -232,23 +236,23 @@ class DeploymentServiceImpl(dispatcher: DeploymentManagerDispatcher,
 
   private def getProcessState(processDetails: BaseProcessDetails[_], inProgressActionTypes: Set[ProcessActionType])
                              (implicit ec: ExecutionContext, freshnessPolicy: DataFreshnessPolicy): DB[ProcessState] = {
-    val manager = dispatcher.deploymentManager(processDetails.processingType)
-    if (inProgressActionTypes.contains(ProcessActionType.Deploy)) {
-      logger.debug(s"Status for: '${processDetails.name}' is: ${SimpleStateStatus.DuringCancel}")
-      DBIOAction.successful(manager.processStateDefinitionManager.processState(SimpleStateStatus.DuringDeploy))
-    } else if (inProgressActionTypes.contains(ProcessActionType.Cancel)) {
-      logger.debug(s"Status for: '${processDetails.name}' is: ${SimpleStateStatus.DuringCancel}")
-      DBIOAction.successful(manager.processStateDefinitionManager.processState(SimpleStateStatus.DuringCancel))
-    } else {
-      checkStateInDeploymentManager(processDetails)
-    }
+    dispatcher.deploymentManager(processDetails.processingType).map { manager =>
+      if (inProgressActionTypes.contains(ProcessActionType.Deploy)) {
+        logger.debug(s"Status for: '${processDetails.name}' is: ${SimpleStateStatus.DuringCancel}")
+        DBIOAction.successful(manager.processStateDefinitionManager.processState(SimpleStateStatus.DuringDeploy))
+      } else if (inProgressActionTypes.contains(ProcessActionType.Cancel)) {
+        logger.debug(s"Status for: '${processDetails.name}' is: ${SimpleStateStatus.DuringCancel}")
+        DBIOAction.successful(manager.processStateDefinitionManager.processState(SimpleStateStatus.DuringCancel))
+      } else {
+        checkStateInDeploymentManager(manager, processDetails)
+      }
+    }.getOrElse(DBIOAction.successful(SimpleProcessStateDefinitionManager.errorFailedToGet))
   }
 
-  private def checkStateInDeploymentManager(processDetails: BaseProcessDetails[_])
+  private def checkStateInDeploymentManager(deploymentManager: DeploymentManager, processDetails: BaseProcessDetails[_])
                                            (implicit ec: ExecutionContext, freshnessPolicy: DataFreshnessPolicy): DB[ProcessState] = {
-    val manager = dispatcher.deploymentManager(processDetails.processingType)
     for {
-      state <- DBIOAction.from(getStateFromEngine(manager, processDetails.idWithName))
+      state <- DBIOAction.from(getStateFromEngine(deploymentManager, processDetails.idWithName))
       cancelActionOpt <- handleFinishedProcess(processDetails, state.value)
     } yield {
       val lastAction = cancelActionOpt.orElse(processDetails.lastAction)
@@ -266,14 +270,28 @@ class DeploymentServiceImpl(dispatcher: DeploymentManagerDispatcher,
   }
 
   private def getStateFromEngine(deploymentManager: DeploymentManager, processIdWithName: ProcessIdWithName)
-                                (implicit ec: ExecutionContext, freshnessPolicy: DataFreshnessPolicy): Future[WithDataFreshnessStatus[Option[ProcessState]]] =
-    deploymentManager.getProcessState(processIdWithName.name).recover {
+                                (implicit ec: ExecutionContext, freshnessPolicy: DataFreshnessPolicy): Future[WithDataFreshnessStatus[Option[ProcessState]]] = {
+    lazy val failedToGetResult =
+      WithDataFreshnessStatus(
+        Option(SimpleProcessStateDefinitionManager.errorFailedToGet),
+        cached = false)
+
+    val stateFromEngineFuture = deploymentManager.getProcessState(processIdWithName.name).recover {
       case NonFatal(e) =>
-        logger.warn(s"Failed to get status of ${processIdWithName}: ${e.getMessage}", e)
-        WithDataFreshnessStatus(
-          Some(SimpleProcessStateDefinitionManager.errorFailedToGet),
-          cached = false)
+        logger.warn(s"Failed to get status of ${processIdWithName.name}: ${e.getMessage}", e)
+        failedToGetResult
     }
+
+    scenarioStateTimeout.map { timeout =>
+      stateFromEngineFuture.withTimeout(timeout, timeoutResult = failedToGetResult).map {
+        case CompletedNormally(value) =>
+          value
+        case CompletedByTimeout(value) =>
+          logger.warn(s"Timeout: $timeout occurred during waiting for response from engine for ${processIdWithName.name}")
+          value
+      }
+    }.getOrElse(stateFromEngineFuture)
+  }
 
   //TODO: there is small problem here: if no one invokes process status for long time, Flink can remove process from history
   //- then it's gone, not finished.