+++ utils/test-utils/src/main/scala/pl/touk/nussknacker/test/VeryPatientScalaFutures.scala
@@ -1,9 +1,8 @@
 package pl.touk.nussknacker.test
 
-import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.time.{Millis, Seconds, Span}
 
-trait VeryPatientScalaFutures extends ScalaFutures with Eventually {
+trait VeryPatientScalaFutures extends BasePatientScalaFutures {
 
   final override implicit def patienceConfig: PatienceConfig = PatienceConfig(timeout = scaled(Span(30, Seconds)), interval = scaled(Span(100, Millis)))